Update client for fleet node authentication
Allow caller to pass `fleet: true` in the initialization parameters to
indicate that we want to use the fleetkey header instead of apikey.

Default to using header authentication instead of query parameters.
Header auth requires rXg version 11.442 or later

Add the `search` method, which performs the same function as the `list`
method but with a filter to restrict the list of results