Improve client library to handle other base_uris
Allow caller to set the base URI for requests rather than assuming it to
be /admin/scaffolds.   This allows the library to be used against other
portions of the rXg's API beyond the scaffold controllers, such as the
/admin/api/fleet controller.

Provide a `with_base_uri` method which accepts a temporary base_uri and
a block to call, and resets the base_uri to its previous value after the
block completes.

Fix a bug that causes inadvertent overriding of class level HTTParty
configuration when instantiating multiple RxgClient instances for
different hosts

Implement basic http action verb utility methods